LUCAS/django/requirements.txt
Django==4.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
Pillow
bcrypt
requests
django-cors-headers
channels
uvicorn
websockets
django-chartjs
daphne

